check if docker is active move autobahn to a new file
