+++ test/js/web/websocket/websocket.test.js
@@ -1,9 +1,7 @@
-import { describe, it, expect, beforeAll, afterAll } from "bun:test";
-import { unsafe, spawn, readableStreamToText, which } from "bun";
-import { tempDirWithFiles, bunExe, bunEnv, gc } from "harness";
+import { describe, it, expect } from "bun:test";
+import { bunExe, bunEnv, gc } from "harness";
 import { readFileSync } from "fs";
 import { join } from "path";
-import child_process from "child_process";
 
 const TEST_WEBSOCKET_HOST = process.env.TEST_WEBSOCKET_HOST || "wss://ws.postman-echo.com/raw";
 
@@ -524,147 +522,5 @@ describe("websocket in subprocess", () => {
 
     server.stop(true);
     expect(await subprocess.exited).toBe(0);
-  });
-});
-
-describe.if(which("docker"))("autobahn", async () => {
-  const url = "ws://localhost:9001";
-  const agent = encodeURIComponent("bun/1.0.0");
-  let docker = null;
-  const { promise, resolve } = Promise.withResolvers();
-  const CWD = tempDirWithFiles("autobahn", {
-    "fuzzingserver.json": `{
-      "url": "ws://127.0.0.1:9001",
-      "outdir": "./",
-      "cases": ["*"],
-      "exclude-agent-cases": {}
-    }`,
-    "index.json": "{}",
-  });
-
-  docker = child_process.spawn(
-    "docker",
-    [
-      "run",
-      "-t",
-      "--rm",
-      "-v",
-      `${CWD}:/config`,
-      "-v",
-      `${CWD}:/reports`,
-      "-p",
-      "9001:9001",
-      "--name",
-      "fuzzingserver",
-      "crossbario/autobahn-testsuite",
-    ],
-    {
-      cwd: CWD,
-      stdout: "pipe",
-      stderr: "pipe",
-    },
-  );
-
-  let out = "";
-  let pending = true;
-  docker.stdout.on("data", data => {
-    out += data;
-    if (pending) {
-      if (out.indexOf("Autobahn WebSocket") !== -1) {
-        pending = false;
-        resolve(true);
-      }
-    }
-  });
-
-  docker.on("close", code => {
-    if (pending) {
-      pending = false;
-      resolve(false);
-    }
-  });
-  const cases = await promise;
-  if (!cases) {
-    throw new Error("Autobahn WebSocket not detected");
-  }
-
-  function getCaseStatus(testID) {
-    return new Promise((resolve, reject) => {
-      const socket = new WebSocket(`${url}/getCaseStatus?case=${testID}&agent=${agent}`);
-      socket.binaryType = "arraybuffer";
-
-      socket.addEventListener("message", event => {
-        resolve(JSON.parse(event.data));
-      });
-      socket.addEventListener("error", event => {
-        reject(event);
-      });
-    });
-  }
-
-  function getTestCaseCount() {
-    return new Promise((resolve, reject) => {
-      const socket = new WebSocket(`${url}/getCaseCount`);
-      let count = null;
-      socket.addEventListener("message", event => {
-        count = parseInt(event.data, 10);
-      });
-      socket.addEventListener("close", event => {
-        if (!count) {
-          reject("No test count received");
-        }
-        resolve(count);
-      });
-    });
-  }
-
-  function getCaseInfo(testID) {
-    return new Promise((resolve, reject) => {
-      const socket = new WebSocket(`${url}/getCaseInfo?case=${testID}`);
-      socket.binaryType = "arraybuffer";
-
-      socket.addEventListener("message", event => {
-        resolve(JSON.parse(event.data));
-      });
-      socket.addEventListener("error", event => {
-        reject(event);
-      });
-    });
-  }
-
-  function runTestCase(testID) {
-    return new Promise((resolve, reject) => {
-      const socket = new WebSocket(`${url}/runCase?case=${testID}&agent=${agent}`);
-      socket.binaryType = "arraybuffer";
-
-      socket.addEventListener("message", event => {
-        socket.send(event.data);
-      });
-      socket.addEventListener("close", event => {
-        resolve();
-      });
-      socket.addEventListener("error", event => {
-        reject(event);
-      });
-    });
-  }
-
-  const count = await getTestCaseCount();
-  it("should have test cases", () => {
-    expect(count).toBeGreaterThan(0);
-  });
-  for (let i = 1; i <= count; i++) {
-    const info = await getCaseInfo(i);
-    const test = parseInt(info.id.split(".")[0]) > 10 ? it.todo : it;
-    // tests > 10 are compression tests, which are not supported yet
-    test(`Running test case ${info.id}: ${info.description}`, async () => {
-      await runTestCase(i);
-      const result = await getCaseStatus(i);
-      expect(["OK", "INFORMATIONAL", "NON-STRICT"]).toContain(result.behavior);
-    });
-  }
-
-  afterAll(() => {
-    docker?.kill();
   });
 });
